refactor: Centralize log listener server handling
