--- packages/aws-lambda-otel-extension/opt/otel-extension/external/lambda-apis/logs-api.js
@@ -1,33 +1,0 @@
-'use strict';
-
-const fetch = require('node-fetch');
-
-const baseUrl = `http://${process.env.AWS_LAMBDA_RUNTIME_API}/2020-08-15/logs`;
-
-async function subscribe(extensionId, subscriptionBody) {
-  const res = await fetch(baseUrl, {
-    method: 'put',
-    body: JSON.stringify(subscriptionBody),
-    headers: {
-      'Content-Type': 'application/json',
-      'Lambda-Extension-Identifier': extensionId,
-    },
-  });
-
-  switch (res.status) {
-    case 200:
-      break;
-    case 202:
-      console.warn(
-        'WARNING!!! Logs API is not supported! Is this extension running in a local sandbox?'
-      );
-      break;
-    default:
-      console.error('logs subscription failed: ', await res.text());
-      break;
-  }
-}
-
-module.exports = {
-  subscribe,
-};
